+++ webtack/jsinterop-generator/src/main/java/org/realityforge/webtack/jsinterop/JsinteropAction.java
@@ -4031,7 +4031,11 @@ final class JsinteropAction
                                          @Nonnull final TypeSpec.Builder type,
                                          @Nonnull final TypeSpec.Builder testType )
   {
-    final boolean requireOverlay = OutputType.j2cl == _outputType && !operation.getReturnType().equals( returnType );
+    final boolean returnTypeNeedsCast =
+      OutputType.j2cl == _outputType && !operation.getReturnType().equals( returnType );
+    final boolean varargsMethod =
+      !typeList.isEmpty() && typeList.get( typeList.size() - 1 ).getJavaType() instanceof ArrayTypeName;
+    final boolean requireOverlay = returnTypeNeedsCast || varargsMethod;
 
     // Codegen does not currently support requiring overlays on java interfaces.
     // We could add that later if ever needed
@@ -4097,7 +4101,7 @@ final class JsinteropAction
     }
     final Type actualReturnType = operation.getReturnType();
     final Kind kind = actualReturnType.getKind();
-    if ( requireOverlay )
+    if ( returnTypeNeedsCast )
     {
       if ( Kind.TypeReference == kind )
       {
@@ -4119,6 +4123,7 @@ final class JsinteropAction
     testCallArgs.add( methodName );
 
     int index = 0;
+    final int argumentCount = arguments.size();
     for ( final Argument argument : arguments )
     {
       if ( 0 == index )
@@ -4137,9 +4142,23 @@ final class JsinteropAction
       testCallStatement.append( "$N" );
       testCallArgs.add( javaName( argument ) );
       final TypedValue typedValue = typeList.get( index++ );
-      generateArgument( argument, typedValue, false, overlayMethod );
+      generateArgument( argument, typedValue, true, overlayMethod );
       generateArgument( argument, typedValue, false, false, method, testMethod );
-    }
+
+      if ( argumentCount == index && varargsMethod )
+      {
+        final TypeName javaType = typedValue.getJavaType();
+        overlayMethod.varargs();
+        final ArrayTypeName javaArrayType = (ArrayTypeName) javaType;
+        if ( javaArrayType.componentType instanceof ArrayTypeName ||
+             javaArrayType.componentType instanceof ParameterizedTypeName )
+        {
+          method.addAnnotation( AnnotationSpec
+                                  .builder( SuppressWarnings.class )
+                                  .addMember( "value", "$S", "unchecked" )
+                                  .build() );
+        }
+      }    }
 
     if ( !arguments.isEmpty() )
     {
@@ -4149,7 +4168,7 @@ final class JsinteropAction
     overlayCallStatement.append( ")" );
     testCallStatement.append( ")" );
 
-    if ( requireOverlay && Kind.TypeReference == kind )
+    if ( returnTypeNeedsCast && Kind.TypeReference == kind )
     {
       overlayCallStatement.append( " )" );
     }