+++ webtack/jsinterop-generator/src/test/fixtures/interface_union_argument/output/j2cl/main/java/com/example/IDBObjectStore.java
@@ -2,7 +2,9 @@ package com.example;
 
 import javax.annotation.Generated;
 import javax.annotation.Nonnull;
+import jsinterop.annotations.JsMethod;
 import jsinterop.annotations.JsNonNull;
+import jsinterop.annotations.JsOverlay;
 import jsinterop.annotations.JsPackage;
 import jsinterop.annotations.JsProperty;
 import jsinterop.annotations.JsType;
@@ -40,6 +42,15 @@ public class IDBObjectStore {
   @JsNonNull
   public native IDBIndex createIndex(@Nonnull String name, @Nonnull JsArray<String> keyPath);
 
+  @JsOverlay
   @JsNonNull
-  public native IDBIndex createIndex(@Nonnull String name, @Nonnull String[] keyPath);
+  public final IDBIndex createIndex(@Nonnull final String name, @Nonnull final String... keyPath) {
+    return _createIndex( name, keyPath );
+  }
+
+  @JsMethod(
+      name = "createIndex"
+  )
+  @JsNonNull
+  private native IDBIndex _createIndex(@Nonnull String name, @Nonnull String[] keyPath);
 }