+++ webtack/jsinterop-generator/src/test/fixtures/transferable/output/gwt/main/java/com/example/MessagePort.java
@@ -4,6 +4,8 @@ import javaemul.internal.annotations.DoNotAutobox;
 import javax.annotation.Generated;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
+import jsinterop.annotations.JsMethod;
+import jsinterop.annotations.JsOverlay;
 import jsinterop.annotations.JsPackage;
 import jsinterop.annotations.JsType;
 
@@ -22,7 +24,16 @@ public class MessagePort {
   public native void postMessage(@DoNotAutobox @Nullable java.lang.Object message,
       @Nonnull JsArray<Object> transfer);
 
-  public native void postMessage(@DoNotAutobox @Nullable java.lang.Object message,
+  @JsOverlay
+  public final void postMessage(@DoNotAutobox @Nullable final java.lang.Object message,
+      @Nonnull final Object... transfer) {
+    _postMessage( message, transfer );
+  }
+
+  @JsMethod(
+      name = "postMessage"
+  )
+  private native void _postMessage(@DoNotAutobox @Nullable java.lang.Object message,
       @Nonnull Object[] transfer);
 
   public native void start();