+++ webtack/jsinterop-generator/src/test/fixtures/type_override/output/j2cl/main/java/com/example/Document.java
@@ -22,7 +22,7 @@ public class Document {
 
   @JsOverlay
   @JsNonNull
-  public final Document open(@Nonnull String unused1, @Nonnull String unused2) {
+  public final Document open(@Nonnull final String unused1, @Nonnull final String unused2) {
     return Js.uncheckedCast( _open( unused1, unused2 ) );
   }
 
@@ -34,7 +34,7 @@ public class Document {
 
   @JsOverlay
   @JsNonNull
-  public final Document open(@Nonnull String unused1) {
+  public final Document open(@Nonnull final String unused1) {
     return Js.uncheckedCast( _open( unused1 ) );
   }
 
@@ -58,7 +58,8 @@ public class Document {
 
   @JsOverlay
   @JsNullable
-  public final Window open(@Nonnull String url, @Nonnull String name, @Nonnull String features) {
+  public final Window open(@Nonnull final String url, @Nonnull final String name,
+      @Nonnull final String features) {
     return Js.uncheckedCast( _open( url, name, features ) );
   }